Tests: unit refactor
