+++ test/unit/README.md
@@ -1,15 +1,10 @@
 ## Setup
 
-- Install [Node.js](https://nodejs.org/)
-- Execute `npm install`
-
-## Build
-
-- Execute `npm run build-test`
+- Execute `npm i --prefix test` from root folder
 
 ## Run
 
 You can run the unit tests in two environments:
 
-- Node.js: Execute `npm run test`
+- Node.js: Execute `npm run test-unit` from root folder
 - Browser: Execute `npm run dev-test` and call `http://localhost:8080/test/unit/UnitTests.html` (see [How to run things locally](https://threejs.org/docs/#manual/introduction/How-to-run-things-locally))
