+++ test/unit/UnitTests_toRemoveAfterEditorES6Refactoring.html
@@ -11,8 +11,8 @@
   <div id="qunit-fixture"></div>
 
   <script src="../node_modules/qunit/qunit/qunit.js"></script>
-  <script src="qunit-utils.js"></script>
-  <script src="SmartComparer.js"></script>
+  <script src="utils/qunit-utils.js"></script>
+  <script src="utils/SmartComparer.js"></script>
 
 <!-- add sources to test below -->
 <script src="../../build/three.min.js"></script>
