+++ test/unit/src/geometries/CircleGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	CircleBufferGeometry
 } from '../../../../src/geometries/CircleGeometry';
