+++ test/unit/src/geometries/CylinderGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	CylinderBufferGeometry
 } from '../../../../src/geometries/CylinderGeometry';
