+++ test/unit/src/geometries/OctahedronGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	OctahedronBufferGeometry
 } from '../../../../src/geometries/OctahedronGeometry';
