+++ test/unit/src/geometries/PlaneGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	PlaneBufferGeometry
 } from '../../../../src/geometries/PlaneGeometry';
