+++ test/unit/src/geometries/PolyhedronGeometry.tests.js
@@ -3,7 +3,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	PolyhedronBufferGeometry
 } from '../../../../src/geometries/PolyhedronGeometry';
