+++ test/unit/src/geometries/RingGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	RingBufferGeometry
 } from '../../../../src/geometries/RingGeometry';
