+++ test/unit/src/geometries/TorusGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	TorusBufferGeometry
 } from '../../../../src/geometries/TorusGeometry';
