+++ test/unit/src/geometries/TorusKnotGeometry.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import {
 	TorusKnotBufferGeometry
 } from '../../../../src/geometries/TorusKnotGeometry';
