+++ test/unit/src/helpers/BoxHelper.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdGeometryTests } from '../../qunit-utils';
+import { runStdGeometryTests } from '../../utils/qunit-utils';
 import { BoxHelper } from '../../../../src/helpers/BoxHelper';
 import { BoxGeometry } from '../../../../src/geometries/BoxGeometry';
 import { SphereGeometry } from '../../../../src/geometries/SphereGeometry';
