+++ test/unit/src/lights/HemisphereLight.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdLightTests } from '../../qunit-utils';
+import { runStdLightTests } from '../../utils/qunit-utils';
 import { HemisphereLight } from '../../../../src/lights/HemisphereLight';
 
 export default QUnit.module( 'Lights', () => {
