+++ test/unit/src/lights/RectAreaLight.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdLightTests } from '../../qunit-utils';
+import { runStdLightTests } from '../../utils/qunit-utils';
 import { RectAreaLight } from '../../../../src/lights/RectAreaLight';
 
 export default QUnit.module( 'Lights', () => {
