+++ test/unit/src/lights/SpotLight.tests.js
@@ -4,7 +4,7 @@
  */
 /* global QUnit */
 
-import { runStdLightTests } from '../../qunit-utils';
+import { runStdLightTests } from '../../utils/qunit-utils';
 import { SpotLight } from '../../../../src/lights/SpotLight';
 
 export default QUnit.module( 'Lights', () => {
